Version 5.12 with NitroFS resource support (old)
This is my custom version of 5.12 from last year, it was based on
version 5.12 (not 5.2). Fixes some bugs and adds support for NitroFS
sprites and backgrounds. Read the full list here:
http://dsgamemaker.com/dsgmforum/viewtopic.php?f=10&t=3174&start=0

+++ SessionsLib.vb
@@ -40,16 +40,21 @@ Module SessionsLib
         File.WriteAllBytes(CompilePath + "gfx\PAGfx.exe", My.Resources.PAGfx)
         Directory.CreateDirectory(CompilePath + "include")
         File.WriteAllBytes(CompilePath + "include\ActionWorks.h", My.Resources.ActionWorks)
-        File.WriteAllBytes(CompilePath + "include\ExtraDBAS.h", My.Resources.ExtraDBAS)
-        File.WriteAllText(CompilePath + "include\ActionWorks.h", PathToString(CompilePath + "include\ActionWorks.h") + vbCrLf + PathToString(CompilePath + "include\ExtraDBAS.h"))
-        File.Delete(CompilePath + "include\ExtraDBAS.h")
+        If IsPro Then
+            File.WriteAllBytes(CompilePath + "include\ExtraDBAS.h", My.Resources.ExtraDBAS)
+            File.WriteAllText(CompilePath + "include\ActionWorks.h", PathToString(CompilePath + "include\ActionWorks.h") + vbcrlf + PathToString(CompilePath + "include\ExtraDBAS.h"))
+            File.Delete(CompilePath + "include\ExtraDBAS.h")
+        End If
         Directory.CreateDirectory(CompilePath + "source")
-        Dim Writeable As String = "make" + vbCrLf + "pause"
+        Dim Writeable As String = "make" + vbcrlf + "pause"
         File.WriteAllText(CompilePath + "build.bat", Writeable)
         Writeable = "make clean"
         File.WriteAllText(CompilePath + "clean.bat", Writeable)
         File.CreateText(CompilePath + "source\main.c").Dispose()
         File.CreateText(CompilePath + "gfx\dsgm_gfx.h").Dispose()
+
+        File.Copy(AppPath + "logo.bmp", CompilePath + "logo.bmp")
+
     End Sub
 
 End Module
